+++ src/longears.c
@@ -3,6 +3,8 @@
 static const R_CallMethodDef longears_entries[] = {
   {"R_amqp_connect", (DL_FUNC) &R_amqp_connect, 6},
   {"R_amqp_is_connected", (DL_FUNC) &R_amqp_is_connected, 1},
+  {"R_amqp_client_properties", (DL_FUNC) &R_amqp_client_properties, 1},
+  {"R_amqp_server_properties", (DL_FUNC) &R_amqp_server_properties, 1},
   {"R_amqp_reconnect", (DL_FUNC) &R_amqp_reconnect, 1},
   {"R_amqp_disconnect", (DL_FUNC) &R_amqp_disconnect, 1},
   {"R_amqp_declare_exchange", (DL_FUNC) &R_amqp_declare_exchange, 8},